requirements.txt
ipfabric # IP Fabric SDK, make sure it matches the version of the IP Fabric instance
loguru
pynetbox
python-dotenv
urllib3
